add size imbalance test case

+++ DataProcessor.py
@@ -44,7 +44,7 @@ class DataProcessor:
         else:
             feature, label = self.test_feature[idx], self.test_label[idx]
         if self.data_source == "cifar":
-            feature = feature.reshape(32, 32, 3).astype(np.float32)
+            feature = feature.reshape(32, 32, 3).astype(np.uint8)
         elif self.data_source == "mnist":
             feature = feature.reshape(28, 28, 1).astype(np.float32)
         img = self.transform(feature)
